Level-2-Agents-with-knowledge-and-storage/requirements.txt
# Core Agno framework
agno>=1.5.1

# Database and storage
sqlalchemy>=2.0.0
lancedb>=0.5.0
tantivy>=0.24.0
pylance>=0.1.0

# Environment and API
python-dotenv>=1.0.0
openai>=1.0.0

# Huggingface embeddings and dependencies
sentence-transformers>=2.2.0
transformers>=4.30.0
huggingface-hub>=0.19.0

# PyTorch and ML dependencies
torch>=2.0.0
scipy>=1.8.0
scikit-learn>=1.0.0
Pillow>=8.0.0
regex>=2024.0.0
safetensors>=0.4.3
tokenizers>=0.21.0

# Network and HTTP dependencies
httpx[socks]>=0.28.0
socksio>=1.0.0
requests>=2.25.0

# Additional dependencies
numpy>=1.25.0,<2.0.0
pandas>=1.5.0
joblib>=1.2.0
threadpoolctl>=3.1.0

# Optional: For better performance
accelerate>=0.20.0 